Work list tests added [no ci]
