+++ app/tests/Behat/BaseContext.php
@@ -11,6 +11,8 @@ use App\Tests\Behat\Traits\GeneralAssertTrait;
 use App\Tests\Behat\Traits\UserTrait;
 use App\Tests\Shared\Traits\FixtureTrait;
 use App\Tests\Shared\Traits\ServiceTrait;
+use App\Tests\Tools\Loader\DoctrineFixtureLoader;
+use App\Tests\Traits\NormalizerTrait;
 use Behat\MinkExtension\Context\MinkContext;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpKernel\KernelInterface;
@@ -19,10 +21,10 @@ class BaseContext extends MinkContext
 {
     use BehatFixtureTrait;
     use BrowserTrait;
-
     use DataTableTrait;
     use FixtureTrait;
     use GeneralAssertTrait;
+    use NormalizerTrait;
     use ServiceTrait;
     use UserTrait;
 
@@ -38,5 +40,13 @@ class BaseContext extends MinkContext
     protected function getService(string $id): object
     {
         return $this->getContainer()->get($id);
+    }
+
+    protected function getReference(string $reference): object
+    {
+        /** @var DoctrineFixtureLoader $fixtureLoader */
+        $fixtureLoader = $this->getService(DoctrineFixtureLoader::class);
+
+        return $fixtureLoader->getReference($reference);
     }
 }
