+++ app/tests/Behat/Traits/BrowserTrait.php
@@ -23,13 +23,4 @@ trait BrowserTrait
     {
         $this->getSession()->wait($seconds * 1000);
     }
-
-    /**
-     * @Then /^Debug: HTML content$/
-     */
-    public function iShowHtmlContent(): void
-    {
-        // @phpstan-ignore-next-line
-        echo $this->getSession()->getPage()->getText();
-    }
 }
