+++ app/tests/Behat/features/contract/acquisitionContract/list.feature
@@ -4,7 +4,6 @@ Feature: Acquisition contract list page
   Background:
     Given I load fixtures:
       | userCompany                  |
-      | contract/acquisitionContract |
       | work/work                    |
     And I am logged on
     And I am on "/acquisition-contracts"
@@ -23,6 +22,7 @@ Feature: Acquisition contract list page
 
   Scenario: I check if datatable has expected rows
     Then I should see "Liste des contrats d'acquisition"
+    And I should see "Afficher les filtres"
     And I should see "1 enregistrements trouvés"
     And I should see a table with:
       | Nom                  | Acquéreur        | Ayant droit | Date de signature | Période de droits       | Œuvres          |
